+++ src/components/_classes/nested/NestedComponent.js
@@ -547,6 +547,7 @@ export default class NestedComponent extends Field {
       header: 'single',
       collapsed: this.collapsed,
       [this.nestedKey]: 'single',
+      messageContainer: 'single-scope',
     });
 
     let childPromise = Promise.resolve();
@@ -762,8 +763,11 @@ export default class NestedComponent extends Field {
     );
   }
 
-  validationProcessor({ scope, data, row, instance }, flags) {
+  validationProcessor({ scope, data, row, instance, component }, flags) {
     const { dirty } = flags;
+    if (this.root.hasExtraPages && this.page !== this.root.page) {
+      instance = this.getComponent(component.path);
+    }
     if (!instance) {
       return;
     }