+++ src/components/datagrid/DataGrid.js
@@ -308,15 +308,6 @@ export default class DataGridComponent extends NestedArrayComponent {
     }, false);
   }
 
-  loadRefs(element, refs) {
-    super.loadRefs(element, refs);
-
-    if (refs['messageContainer'] === 'single') {
-      const container = _.last(element.querySelectorAll(`[${this._referenceAttributeName}=messageContainer]`));
-      this.refs['messageContainer'] = container || this.refs['messageContainer'];
-    }
-  }
-
   attach(element) {
     this.loadRefs(element, {
       [`${this.datagridKey}-row`]: 'multiple',
@@ -325,7 +316,6 @@ export default class DataGridComponent extends NestedArrayComponent {
       [`${this.datagridKey}-removeRow`]: 'multiple',
       [`${this.datagridKey}-group-header`]: 'multiple',
       [this.datagridKey]: 'multiple',
-      'messageContainer': 'single'
     });
 
     if (this.allowReorder) {