+++ src/components/editgrid/EditGrid.js
@@ -573,11 +573,7 @@ export default class EditGridComponent extends NestedArrayComponent {
       this.removeClass(this.refs.component, `formio-component-${this.component.type}-row-open`);
     }
 
-    const superAttach = super.attach(element);
-    this.loadRefs(element, {
-      messageContainer: 'single-scope',
-    });
-    return superAttach;
+    return super.attach(element);
   }
 
   flattenRowDataValue(dataValue) {