+++ Changelog.md
@@ -34,6 +34,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - FIO-8627: Fixed the email table generation for nested array components
  - FIO-5748: Fixed preview not working correctly after hiding/showing it
  - FIO-8759: fixed validation for the default value component
+ - FIO-8765: fixed an issue where dataMap thows an error
+ - FIO-8783: Fixed the issue when validation messages disappear after collapse and expand the panel
+ - FIO-8072: Added conditional operators for select component with number data type
+ - FIO-8752: coerced keyboard actions to a boolean value
+ - FIO-8656: lazy load dragula
+ - FIO-8570: fixed thousandSeparator problem and decimal symbol problem
+ - FIO-8710: fixed incorrect Components import that causes errors when interacting with form components
 
 ### New Features