+++ Changelog.md
@@ -42,7 +42,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - FIO-8570: fixed thousandSeparator problem and decimal symbol problem
  - FIO-8710: fixed incorrect Components import that causes errors when interacting with form components
  - FIO-8801: fixed submission metadata nullish
- - FIO-8809: fixed runtime error in setForm when passing string argument to Formio.builder 
+ - FIO-8809: fixed runtime error in setForm when passing string argument to Formio.builder
+ - FIO-8684: Fixes cannot attach more than one component to the PDF form
+ - FIO-8719: fixed validation for Data Grid component inside nested wizard
 
 ### New Features