+++ Changelog.md
@@ -28,6 +28,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - FIO-8633 fixed icon name for Edit Grid
  - FIO-8457: set pristine flag when datagrid reorders rows
  - FIO-8716: fixed unsavedRows validation for Edit Grid
+ - FIO-8647,8721: validation not triggered for each row after the first one in data and edit grid
+ - FIO-8720: fixed errors occurring with valid time and number components
+ - FIO-8719: fixed validation for nested wizard fields
+ - FIO-8627: Fixed the email table generation for nested array components
+ - FIO-5748: Fixed preview not working correctly after hiding/showing it
+ - FIO-8759: fixed validation for the default value component
+ - FIO-8765: fixed an issue where dataMap thows an error
+ - FIO-8783: Fixed the issue when validation messages disappear after collapse and expand the panel
+ - FIO-8072: Added conditional operators for select component with number data type
+ - FIO-8752: coerced keyboard actions to a boolean value
+ - FIO-8656: lazy load dragula
+ - FIO-8570: fixed thousandSeparator problem and decimal symbol problem
+ - FIO-8710: fixed incorrect Components import that causes errors when interacting with form components
+ - FIO-8801: fixed submission metadata nullish
+ - FIO-8809: fixed runtime error in setForm when passing string argument to Formio.builder
+ - FIO-8684: Fixes cannot attach more than one component to the PDF form
+ - FIO-8719: fixed validation for Data Grid component inside nested wizard
 
 ### New Features