+++ Changelog.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased: 5.0.0-rc.72]
+## 5.0.0-rc.72
+### Changed
+ - Updated @formio/core@2.2.2-rc.1
+ - FIO-8684: Fixes cannot attach more than one component to the PDF form
+ - Bugfix/submission metadata nullish
+ - FIO-8765: fixed an issue where dataMap thows an error
+ - FIO-8719 fixed validation for Data Grid component inside nested wizard
+
 ### Changed
  - FIO-8621: error event on file upload adding
  - FIO-8783: Fixed the issue when validation messages disappear after collapse and expand the panel